requirement.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
transformers==4.35.0
torch==2.1.0
Pillow==10.0.1
langchain==0.0.335
gTTS==2.4.0
Flask-Caching==2.1.0
redis==5.0.1